Fix warnings from gcc -m64
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/branches/V5-1-patches@9189 06827809-a52a-0410-b366-d66718629ded

+++ agent/helpers/table_iterator.c
@@ -143,7 +143,7 @@ netsnmp_insert_iterator_context(netsnmp_request_info *request, void *data)
     oid      base_oid[] = {0, 0};	/* Make sure index OIDs are legal! */
     oid      this_oid[MAX_OID_LEN];
     oid      that_oid[MAX_OID_LEN];
-    int      this_oid_len, that_oid_len;
+    size_t   this_oid_len, that_oid_len;
 
     if (!request)
         return;
